backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.0.3
opencv-python==4.8.1.78
pillow==10.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
numpy>=1.21.0
